% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{CreateProjectFolders}
\alias{CreateProjectFolders}
\title{Creates project infrastructure}
\usage{
CreateProjectFolders()
}
\description{
This function creates the folders you will use in your project
}
\examples{
CreateProjectFolders()
}
\keyword{project}
